Moves axios requests out of component components
