+++ src/components/CityDetails.jsx
@@ -1,9 +1,9 @@
 import React from "react";
-import axios from "axios";
 
 import Row from "react-bootstrap/Row";
 import Col from "react-bootstrap/Col";
 import { CityMap, Weather, Movies } from "./";
+import { getWeatherForecast, getTopMovies } from "../assets/axios.js";
 
 export default class CityDetails extends React.Component {
   constructor(props) {
@@ -18,44 +18,11 @@ export default class CityDetails extends React.Component {
   }
 
   componentDidMount = async () => {
+    const { currentCity } = this.props;
     const newState = {};
-    newState.weather = await this.getWeatherForecast();
-    newState.movies = await this.getTopMovies();
+    newState.weather = await getWeatherForecast(currentCity);
+    newState.movies = await getTopMovies(currentCity);
     this.setState(newState);
-  };
-
-  getWeatherForecast = async () => {
-    const { display_name, lat, lon } = this.props.currentCity;
-    const [cityName] = display_name.split(",") || [display_name];
-    const url = `${
-      import.meta.env.VITE_SERVER_URL
-    }/weather?searchQuery=${cityName}&lat=${lat}&lon=${lon}`;
-
-    const weather = {};
-    try {
-      // const { data: forecasts } = await axios.get(url);
-      weather.forecasts = []; // forecasts;
-    } catch ({ response: { statusText: errorMessage } }) {
-      weather.errorMessage = errorMessage;
-    }
-    return weather;
-  };
-
-  getTopMovies = async () => {
-    const { display_name } = this.props.currentCity;
-    const [cityName] = display_name.split(",");
-    const url = `${
-      import.meta.env.VITE_SERVER_URL
-    }/movies?searchQuery=${cityName}`;
-
-    const movies = {};
-    try {
-      const { data: top } = await axios.get(url);
-      movies.top = top;
-    } catch ({ response: { statusText: errorMessage } }) {
-      movies.errorMessage = errorMessage;
-    }
-    return movies;
   };
 
   render() {
@@ -67,14 +34,14 @@ export default class CityDetails extends React.Component {
 
     return (
       <>
-        <Row>
+        <Row className="mb-2">
           <Col>
             <h2>{display_name}</h2>
             <hr />
           </Col>
         </Row>
-        <Row>
-          <Col xs={12} md={6}>
+        <Row className="mb-2">
+          <Col xs={12} lg={6}>
             <CityMap currentCity={currentCity} />
           </Col>
           <Col xs={12} lg={6}>
@@ -85,7 +52,7 @@ export default class CityDetails extends React.Component {
             />
           </Col>
         </Row>
-        <Row>
+        <Row className="mb-2">
           <Col>
             <Movies
               key={`movies-${cityName}`}
